Add functionality to list users in a ride.
Enhanced the ReservationsRepository to include a method for retrieving users in a specific ride. Updated RideService to filter rides by status "in-progress" and modified RideController redirections to consistently point to "/rides/myRides". Adjusted HTML file for improved image sourcing and made changes to security configurations and application properties for better static resource handling.